Better storage for cached entries
A single JSON object is not very efficient to parse as it requires
loading the whole file.

This commit implements saving the entries as json documents, one entry
per line, in descending order of dates (newest first). This way we can
limit ourselves to reading the files line by line while performing a
merge sort of all the files.