Fix wrong feed name in the generated pages
Before this commit, all the entries displayed the same feed name because
the generator was resolving the "feed" variable when running the
generator (as opposed to when defining it). This caused the bound "feed"
variable to contain the latest value it contains during the for loop,
i.e. the latest item in the list of feeds.

To work around this, the generator is now created in another function,
so the "feed" variable within that function doesn't change.